feat(player): migrate 8 clusters to generated types (big batch)
Nineteenth call-site migration (#461). Eight domains in one PR to
cut CI overhead — user profile, preferences, devices, game
sessions, session cheats, game cheats, cores, key mappings.

Unlocked by #479's GameDto typealias; many of these DTOs nest
GameResponse / CollectionResponse / other now-generated types.

## Clusters migrated

- **User profile**: UserDto -> UserResponse (typealiased because
  hand-written AuthResponse.user keeps pointing at the alias —
  auth endpoints haven't been huma-migrated).
- **Preferences**: UserPreferencesDto, UpdatePreferencesRequest,
  ConsoleKeyMappingDto deleted. Generated spec uses
  ConsoleKeyMappingDTO (uppercase) — renamed accordingly. Nullable
  customMapping handled with .orEmpty().
- **Devices**: DeviceDto typealiased to DeviceResponse (used in
  presentation layer). RegisterDeviceRequest +
  UpdateDevicePreferencesRequest deleted. lastSeenAt went from
  String to non-null Instant — removed a vestigial .isNotEmpty()
  branch in DeviceManagementSection that was never hit.
- **Game sessions**: GameSessionDto, CreateSessionRequest,
  UpdateSessionRequest deleted. DuplicateSessionRequest now from
  generated package. Mapper handles memberCount Long->Int,
  lastPlayedAt Instant->String, nullable memberUsernames /
  memberAvatars via .orEmpty().
- **Session cheats**: SessionCheatConfigDto,
  UpdateSessionCheatsRequest deleted. enabledIndices is
  List<Long>? in the spec (huma quirk) — mapper uses
  .orEmpty().map { it.toInt() }; Int<->Long conversion at the
  boundary for update requests.
- **Game cheats**: CheatDto deleted. CheatRepositoryImpl drops
  .toLong() on the now-Long dto.index.
- **Cores**: LibretroCoreDto deleted; mapper on generated Core.
  getRecommendedCore returns domain LibretroCore directly since
  the {coreName:"..."} fallback can't satisfy Core's @Required
  createdAt/updatedAt Instants.
- **Key mappings**: GameKeyMappingDto, UpdateGameKeyMappingRequest
  deleted; unused outside the data layer.

## Cluster skipped

- **Session saves** (SaveStateDto -> SessionSaveResponse): domain
  SaveState.id: Long, gameId: Long but generated has id: String
  and no gameId. Requires a domain-model change across tests,
  UI, and repository — outside this PR's scope. Left hand-written
  for a follow-up.

## Test plan

- [x] Full compile green across :shared / :shared-api / :desktop
- [x] ./run-desktop-tests.sh — 470/470 pass
- [x] Pre-existing SessionsUiTest E2E failures unchanged
- [ ] CI green

Refs #461.

+++ player/shared/src/commonMain/kotlin/com/spela/player/data/repository/CheatRepositoryImpl.kt
@@ -31,7 +31,7 @@ class CheatRepositoryImpl(
                 queries.upsertCheat(
                     id = cheatId,
                     game_id = gameId,
-                    cheat_index = dto.index.toLong(),
+                    cheat_index = dto.index,
                     description = dto.description,
                     code = dto.code,
                     enabled = if (wasEnabled) 1L else 0L,
